feat: member limit for Hera app permissions (#629)
* refactor: extend the id column in the hera_app_role table to the bigint type

* feat: allow configuring platform codes to prohibit MQ messages from accessing the database

* feat: member limit for Hera app permissions

* fix: member limit for Hera app permissions

---------

+++ ozhera-monitor/ozhera-monitor-service/src/main/java/org/apache/ozhera/monitor/dao/HeraAppRoleDao.java
@@ -38,7 +38,7 @@ public class HeraAppRoleDao {
     @Resource
     private HeraAppRoleMapper heraAppRoleMapper;
 
-    public Integer delById(Integer id) {
+    public Integer delById(Long id) {
         if (id == null) {
             log.error("HeraAppRoleDao.delById invalid param,id is null");
         }
